[Refactor] UI 필요없는 함수 삭제
