+++ app/src/main/java/kr/sparta/tripmate/ui/mypage/home/MyPageTabLayoutAdapter.kt
@@ -35,7 +35,4 @@ class MyPageTabLayoutAdapter(fragment: FragmentActivity) : FragmentStateAdapter(
     fun updateScrap(): Fragment {
         return fragments[1].fragment
     }
-    fun updateBoard():Fragment{
-        return fragments[0].fragment
-    }
 }
