+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/scrap/main/SearchBlogFactory.kt
@@ -3,7 +3,7 @@ package kr.sparta.tripmate.ui.viewmodel.scrap.main
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import kr.sparta.tripmate.api.naver.NaverNetWorkClient
-import kr.sparta.tripmate.data.datasource.local.budget.SharedPreferencesLocalDataSource
+import kr.sparta.tripmate.data.datasource.local.sharedpreference.SharedPreferencesLocalDataSource
 import kr.sparta.tripmate.data.datasource.remote.community.scrap.FirebaseBlogScrapRemoteDataSource
 import kr.sparta.tripmate.data.repository.ImageRepositoryImpl
 import kr.sparta.tripmate.data.repository.community.FirebaseBlogScrapRepositoryImpl
